Changes: 1. Change primitive type int for id in entities to Long type 2. Refactor entity tests with Long type 3. Add @JoinTable annotation for account_category table

+++ src/main/java/com/dzvonik/cashflow/domain/entity/Category.java
@@ -2,7 +2,6 @@ package com.dzvonik.cashflow.domain.entity;
 
 import java.util.List;
 
-import javax.persistence.CascadeType;
 import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.Id;
@@ -15,7 +14,6 @@ import javax.persistence.FetchType;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
-import lombok.Builder;
 import lombok.EqualsAndHashCode;
 import lombok.NoArgsConstructor;
 import lombok.ToString;
@@ -31,7 +29,7 @@ public class Category {
     @Id
     @SequenceGenerator(name = "seq_category", sequenceName = "seq_category")
     @GeneratedValue(strategy = GenerationType.SEQUENCE, generator = "seq_category")
-    private int id;
+    private Long id;
 
     @Column(nullable = false)
     private String title;
